Novos ajustes no site do tcc

+++ TccSite.Domain/Entities/Alerta.cs
@@ -29,6 +29,6 @@ namespace TccSite.Domain.Entities
 
         public string Descricao { get; set; }
 
-        public int? NivelRio { get; set; }
+        public decimal? NivelRio { get; set; }
     }
 }
