+++ TccSite/Models/ViewModels/AlertaGraficoViewModel.cs
@@ -4,7 +4,7 @@
     {
         public int CodAlerta { get; set; }
         public DateTime? DataCadastro { get; set; }
-        public int? NivelRio { get; set; }
+        public decimal? NivelRio { get; set; }
         public string StatusAlerta { get; set; }
     }
 }
